automatically end the game when two cards remain
The requirement for a human player to type "end" when two cards
remain to end the game is easy for the player to overlook.  If the
player plays beyond this situation, the game cannot end.

Therefore, automatically end the game when only two cards remain
(they are necessarily matching cards).  The user can still end the
game early by typing "end".

Also add an alert message at the very beginning of the game to
inform the user as to how to (prematurely) end the game.